+++ packer/debian-7.4.0-amd64/scripts/cleanup.sh
@@ -30,9 +30,6 @@ rm -rf /usr/src/vboxguest*
 rm -rf /usr/share/man/??
 rm -rf /usr/share/man/??_*
 
-echo "Adding a 2 sec delay to the interface up, to make the dhclient happy"
-echo "pre-up sleep 2" >> /etc/network/interfaces
-
 # Zero out the free space to save space in the final image:
 echo "Zeroing device to make space..."
 dd if=/dev/zero of=/EMPTY bs=1M