+++ packer/debian-7.4.0-amd64/scripts/vmtools.sh
@@ -10,6 +10,7 @@ virtualbox-iso|virtualbox-ovf)
     umount /tmp/vbox
     rmdir /tmp/vbox
     rm /home/vagrant/*.iso
+    ln -s /opt/VBoxGuestAdditions-*/lib/VBoxGuestAdditions /usr/lib/VBoxGuestAdditions
     ;;
 
 vmware-iso|vmware-ovf)