+++ packer/ubuntu-12.04.4-server-amd64/scripts/cleanup.sh
@@ -30,18 +30,6 @@ rm -rf /usr/src/vboxguest*
 rm -rf /usr/share/man/??
 rm -rf /usr/share/man/??_*
 
-# Removing leftover leases and persistent rules
-echo "cleaning up dhcp leases"
-rm /var/lib/dhcp/*
-
-# Make sure Udev doesn't block our network
-mkdir /etc/udev/rules.d/70-persistent-net.rules
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
-rm -rf /dev/.udev/ /var/lib/dhcp3/*
-
-echo "Adding a 2 sec delay to the interface up, to make the dhclient happy"
-echo "pre-up sleep 2" >> /etc/network/interfaces
-
 # Zero out the free space to save space in the final image:
 echo "Zeroing device to make space..."
 dd if=/dev/zero of=/EMPTY bs=1M